requirements.txt
asgiref==3.8.1
asttokens==3.0.0
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==43.0.3
daphne==4.1.2
decorator==5.2.1
Django==5.0
django-extensions==4.1
django-parler==2.3
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
executing==2.2.1
geographiclib==2.0
geopy==2.4.1
googlemaps==4.10.0
gunicorn==21.2.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
iniconfig==2.1.0
ipython==9.5.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
loguru==0.7.0
matplotlib-inline==0.1.7
packaging==24.1
parso==0.8.5
pexpect==4.9.0
Pillow==10.0.0
pluggy==1.6.0
prompt_toolkit==3.0.52
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.19.2
pyOpenSSL==24.2.1
pytest==8.4.1
pytest-django==4.11.1
python-dotenv==1.0.0
pytz==2024.2
qrcode==8.0
requests==2.32.3
service-identity==24.1.0
sqlparse==0.5.1
stack-data==0.6.3
stripe==10.12.0
traitlets==5.14.3
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
Unidecode==1.4.0
urllib3==2.2.3
wcwidth==0.2.13
whitenoise==6.6.0
zope.interface==7.1.0
psycopg2-binary==2.9.9
